[#3]Feat: 찜하기 기능 구현

+++ src/main/java/com/inno/coogle/dto/post/PostRequestDto.java
@@ -10,7 +10,7 @@ public class PostRequestDto {
     private String postContents;
     private List<String> ingredientsList;
     private List<String> tagList;
-    private Long level;
+    private int level;
     private String foodType;
 
 }
